+++ app/src/main/java/com/example/patrick/grocr2/PlayGround.java
@@ -12,7 +12,7 @@ import java.util.ArrayList;
 public class PlayGround {
 
     public static void main(String[] args) {
-        Orders order = new Orders(1.,1.,"s",true,0,0,0,new ArrayList<Float>());
+        Orders order = new Orders(1.,1.,"s",true,0,0,0,new ArrayList<Long>());
 
         String serializedObject = "";
